reducing repetition in info and entry classes
