+++ TBI-GUI/src/utils/InfoManagement.java
@@ -1,0 +1,5 @@
+package utils;
+
+public class InfoManagement {
+
+}
