+++ TBI-GUI/src/utils/PatientEntry.java
@@ -4,42 +4,14 @@ import java.io.Serializable;
 import java.security.Key;
 import java.util.Date;
 
-public class PatientEntry implements Serializable, Comparable{
-	public String name;
-	public String uid;
-	public Key key;
-	public Date date;
+public class PatientEntry extends InfoEntry implements Comparable, Serializable{
 	
 	public PatientEntry(String n, String u, Key k) {
-		name = n;
-		uid = u;
-		key = k;
-		date = new Date();
-	}
-
-	public final String getName() {
-		return name;
-	}
-
-	public final String getUid() {
-		return uid;
-	}
-	
-	public final Date getLastDate() {
-		return date;
+		super(n, u, k);
 	}
 
 	public int compareTo(Object o) {
-		PatientEntry other = (PatientEntry)o;
-		int comp = date.compareTo(other.date);
-		if(comp == 0) {
-			comp = name.compareTo(other.name);
-			if(comp == 0) {
-				comp = uid.compareTo(other.uid);
-			}
-		}
-		
-		return comp;
+		return super.compareTo(o);
 	}
 	
 }
